pyproject.toml
[tool.poetry]
name = "netban-core"
version = "0.0.1"
description = "Netban - a tool for banning users per 'networks' of channels on a social media."
authors = ["moonburnt <moonburnt@disroot.org>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-postgres-extra = "^2.0.8"
psycopg2 = "^2.9.10"
drf-spectacular = "^0.28.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"

[tool.ruff]
indent-width = 4
line-length = 80
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint]
select = [
	"E4", 
	"E7", 
	"E9", 
	"F"
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

